porting libcvd to Win32: - added VC2005 build files   - these create a static libcvd library (to avoid dllimport/dllexport hassle)   - the library is created in release and debug version   - runtime library is the static one /MT or /MTd for debug   - the configuration directory contains a Win32 specific config.h header file which is copied by the build system to cvd/ before compilation. make any changes there - a new OS specific cvd_src/Win32 directory contains platform specific implementations of GLWindow and aligned memory allocation and some helpers - aligned mem has now a unified interface and at build time switches between the buildin implementation and posix_memalign and Win32 aligned allocation. - the port depends on the pthreads-win32 library http://sourceware.org/pthreads-win32/ and libjpeg. other image libraries should be usable as well, need to be added to config.h and build system
I didn't make any attempt to come up with a solution to specifying where other libraries and header files are. Use the project settings to configure these.

+++ cvd/internal/builtin_components.h
@@ -52,14 +52,14 @@ namespace CVD
     template<class P> struct Component
     {
       typedef P type;
-      static const unsigned int count = 1;
+      static const size_t count = 1;
 		  
-      static const P& get(const P& pixel, int)
+      static const P& get(const P& pixel, size_t)
       {
 	return pixel;
       }
 
-      static P& get(P& pixel, int)
+      static P& get(P& pixel, size_t)
       {
 	return pixel;
       }
@@ -69,14 +69,14 @@ namespace CVD
     template<class P, int I> struct Component<P[I]>
     {
       typedef P type;
-      static const unsigned int count=I;
+      static const size_t count=I;
       
-      static const P& get(const P pixel[I], unsigned int i)
+      static const P& get(const P pixel[I], size_t i)
       {
 	return pixel[i];
       }
 
-      static inline P& get(P pixel[I], unsigned int i)
+      static inline P& get(P pixel[I], size_t i)
       {
 	return pixel[i];
       }
@@ -86,14 +86,14 @@ namespace CVD
     template<int N> struct Component<TooN::Vector<N> >
     {
       typedef double type;
-      static const unsigned int count=N;
+      static const size_t count=N;
       
-      static const type& get(const TooN::Vector<N>& pixel, unsigned int i)
+      static const type& get(const TooN::Vector<N>& pixel, size_t i)
       {
 	return pixel[i];
       }
 
-      static inline type& get(TooN::Vector<N>& pixel, unsigned int i)
+      static inline type& get(TooN::Vector<N>& pixel, size_t i)
       {
 	return pixel[i];
       }
